trivial: Rename algo/crow to algo/minotaurx
