+++ src/algo/minotaurx/blake2s-ref.c
