+++ src/algo/minotaurx/hamsi_helper.c
