+++ src/algo/minotaurx/hashquark.h
