+++ src/algo/minotaurx/hashqubit.h
