+++ src/algo/minotaurx/hashskein.h
