+++ src/algo/minotaurx/hashx15.h
