+++ src/algo/minotaurx/haval_helper.c
