+++ src/algo/minotaurx/md_helper.c
