+++ src/algo/minotaurx/yespower/README
