+++ src/algo/minotaurx/yespower/util/sysendian.h
