+++ src/algo/minotaurx/yespower/yespower.c
