+++ src/algo/minotaurx/yespower/yespower.h
